feat: create @elizaos/utils package for shared build utilities
- Create new @elizaos/utils package with standardized build utilities
- Move build utilities from root build-utils.ts to @elizaos/utils package
- Update all packages to import from @elizaos/utils instead of build-utils
- Add @elizaos/utils as workspace dependency to all relevant packages
- Remove old build-utils.ts file from root directory

This change improves code organization by properly packaging build utilities
and eliminates the need for relative imports across the monorepo.
