+++ packages/api-client/package.json
@@ -15,7 +15,8 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@elizaos/core": "workspace:*"
+    "@elizaos/core": "workspace:*",
+    "@elizaos/utils": "workspace:*"
   },
   "devDependencies": {
     "@types/bun": "latest",
