+++ packages/cli/package.json
@@ -85,6 +85,7 @@
     "@elizaos/plugin-sql": "workspace:*",
     "@elizaos/server": "workspace:*",
     "@elizaos/plugin-bootstrap": "workspace:*",
+    "@elizaos/utils": "workspace:*",
     "bun": "^1.2.17",
     "chalk": "^5.4.1",
     "chokidar": "^4.0.3",
