+++ packages/config/package.json
@@ -37,6 +37,9 @@
     "dist",
     "src"
   ],
+  "dependencies": {
+    "@elizaos/utils": "workspace:*"
+  },
   "devDependencies": {
     "prettier": "^3.5.3",
     "typescript": "^5.8.2"
