+++ packages/core/package.json
@@ -42,6 +42,7 @@
     "typescript": "5.8.3"
   },
   "dependencies": {
+    "@elizaos/utils": "workspace:*",
     "@sentry/browser": "^9.22.0",
     "buffer": "^6.0.3",
     "crypto-browserify": "^3.12.1",
