+++ packages/plugin-bootstrap/package.json
@@ -27,6 +27,7 @@
   "dependencies": {
     "@elizaos/core": "workspace:*",
     "@elizaos/plugin-sql": "workspace:*",
+    "@elizaos/utils": "workspace:*",
     "bun": "^1.2.17"
   },
   "devDependencies": {
