+++ packages/plugin-sql/package.json
@@ -28,6 +28,7 @@
   "dependencies": {
     "@electric-sql/pglite": "^0.3.3",
     "@elizaos/core": "workspace:*",
+    "@elizaos/utils": "workspace:*",
     "dotenv": "^16.4.7",
     "drizzle-kit": "^0.31.1",
     "drizzle-orm": "^0.44.2",
