+++ packages/project-starter/package.json
@@ -30,6 +30,7 @@
   "dependencies": {
     "@elizaos/cli": "workspace:*",
     "@elizaos/core": "workspace:*",
+    "@elizaos/utils": "workspace:*",
     "@elizaos/plugin-bootstrap": "workspace:*",
     "@elizaos/plugin-sql": "workspace:*",
     "@tanstack/react-query": "^5.29.0",
