+++ packages/project-tee-starter/package.json
@@ -35,6 +35,7 @@
   "dependencies": {
     "@elizaos/cli": "workspace:*",
     "@elizaos/core": "workspace:*",
+    "@elizaos/utils": "workspace:*",
     "@elizaos/plugin-bootstrap": "workspace:*",
     "@elizaos/plugin-redpill": "1.2.1",
     "@elizaos/plugin-sql": "workspace:*",
