+++ packages/test-utils/package.json
@@ -39,6 +39,7 @@
   ],
   "dependencies": {
     "@elizaos/core": "workspace:*",
+    "@elizaos/utils": "workspace:*",
     "zod": "3.24.2"
   },
   "devDependencies": {
